Move logic from profile360 to profile
