+++ src/applications/personalization/profile/manifest.json
