+++ src/applications/personalization/profile/reducers/index.js
