+++ src/applications/personalization/profile/util/connected-apps.js
