+++ src/applications/personalization/profile360/components/BankInfoForm.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import SchemaForm from 'platform/forms-system/src/js/components/SchemaForm';
 import LoadingButton from 'platform/site-wide/loading-button/LoadingButton';
 
-import { ACCOUNT_TYPES_OPTIONS } from '../constants';
+import { ACCOUNT_TYPES_OPTIONS } from 'applications/personalization/profile/constants.js';
 
 export const schema = {
   type: 'object',
